restructure dependency
